+++ src/gracenotegroup.ts
@@ -6,7 +6,7 @@
 // render grace notes.
 
 import { log } from './util';
-import { Flow } from './tables';
+import { Flow } from './flow';
 import { Modifier } from './modifier';
 import { Formatter } from './formatter';
 import { Voice } from './voice';
